% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/da_Counting_Replicates_per_Time.R
\name{da_Counting_Replicates_per_Time}
\alias{da_Counting_Replicates_per_Time}
\title{Auxiliary function for MacKenzie mle methods}
\usage{
da_Counting_Replicates_per_Time(Time)
}
\arguments{
\item{Time}{an array containing sampling times (including time repetitions)

The output of this function generates a list containing a time vector 'Time_Vector'
with all sampling times (without repetitions) and an integer vector 'Transects'
counting the number of repetitions or replicates per sampling time.}
}
\description{
\code{da_Counting_Replicates_per_Time} counts transects/replicates per sampling time.
The number of replicates per sampling time can perfectly differ.
}
\examples{
Time <- c(2000,2000,2000,2001,2001,2002,2002,2002,2002)
R <- da_Counting_Replicates_per_Time( Time )
}
\keyword{MacKenzie,}
\keyword{colonization,}
\keyword{detectability}
\keyword{extinction,}
